extract sql logic into vegafusion-sql crate
