+++ vegafusion-datafusion-udfs/src/udfs/array/mod.rs
@@ -1,3 +1,4 @@
 pub mod indexof;
 pub mod length;
 pub mod span;
+pub mod constructor;
