+++ vegafusion-datafusion-udfs/src/udfs/datetime/mod.rs
@@ -7,6 +7,7 @@ pub mod str_to_timestamptz;
 pub mod timestamp_to_timestamptz;
 pub mod timestamptz_to_epoch;
 pub mod timestamptz_to_timestamp;
+pub mod timeunit;
 
 use crate::udfs::datetime::str_to_timestamptz::datetime_strs_to_timestamp_millis;
 use std::sync::Arc;
