+++ vegafusion-runtime/src/lib.rs
@@ -1,15 +1,11 @@
 #[macro_use]
 extern crate lazy_static;
-
-#[macro_use]
-extern crate log;
 extern crate core;
 
 pub mod data;
 pub mod expression;
 pub mod pre_transform;
 pub mod signal;
-pub mod sql;
 pub mod task_graph;
 pub mod tokio_runtime;
 pub mod transform;
