--- vegafusion-runtime/src/sql/compile/mod.rs
@@ -1,6 +1,0 @@
-pub mod data_type;
-pub mod expr;
-pub mod function_arg;
-pub mod order;
-pub mod scalar;
-pub mod select;
