--- vegafusion-runtime/src/sql/mod.rs
@@ -1,4 +1,0 @@
-pub mod compile;
-pub mod connection;
-pub mod dataframe;
-pub mod dialect;
