+++ vegafusion-runtime/src/task_graph/task.rs
@@ -1,5 +1,4 @@
 use crate::data::dataset::VegaFusionDataset;
-use crate::sql::connection::Connection;
 use crate::task_graph::timezone::RuntimeTzConfig;
 use async_trait::async_trait;
 use std::collections::HashMap;
@@ -9,6 +8,7 @@ use vegafusion_core::error::Result;
 use vegafusion_core::proto::gen::tasks::task::TaskKind;
 use vegafusion_core::proto::gen::tasks::Task;
 use vegafusion_core::task_graph::task_value::TaskValue;
+use vegafusion_dataframe::connection::Connection;
 
 #[async_trait]
 pub trait TaskCall {
