+++ vegafusion-runtime/src/transform/collect.rs
@@ -7,7 +7,6 @@ use std::sync::Arc;
 use vegafusion_core::error::{Result, ResultWithContext};
 use vegafusion_core::proto::gen::transforms::{Collect, SortOrder};
 
-use crate::sql::dataframe::DataFrame;
 use async_trait::async_trait;
 use datafusion::common::ScalarValue;
 use datafusion_expr::{
@@ -16,6 +15,7 @@ use datafusion_expr::{
 use vegafusion_common::column::{flat_col, unescaped_col};
 use vegafusion_common::data::ORDER_COL;
 use vegafusion_core::task_graph::task_value::TaskValue;
+use vegafusion_dataframe::dataframe::DataFrame;
 
 #[async_trait]
 impl TransformTrait for Collect {
