+++ vegafusion-runtime/src/transform/fold.rs
@@ -1,7 +1,6 @@
 use crate::expression::compiler::config::CompilationConfig;
 use crate::transform::TransformTrait;
 
-use crate::sql::dataframe::DataFrame;
 use async_trait::async_trait;
 use std::sync::Arc;
 use vegafusion_common::data::ORDER_COL;
@@ -9,6 +8,7 @@ use vegafusion_common::error::Result;
 use vegafusion_common::escape::unescape_field;
 use vegafusion_core::proto::gen::transforms::Fold;
 use vegafusion_core::task_graph::task_value::TaskValue;
+use vegafusion_dataframe::dataframe::DataFrame;
 
 #[async_trait]
 impl TransformTrait for Fold {
