+++ vegafusion-runtime/src/transform/identifier.rs
@@ -1,7 +1,6 @@
 use crate::expression::compiler::config::CompilationConfig;
 use crate::transform::TransformTrait;
 
-use crate::sql::dataframe::DataFrame;
 use async_trait::async_trait;
 use datafusion_expr::{
     expr, BuiltInWindowFunction, Expr, WindowFrame, WindowFrameBound, WindowFrameUnits,
@@ -14,6 +13,7 @@ use vegafusion_common::data::ORDER_COL;
 use vegafusion_common::error::Result;
 use vegafusion_core::proto::gen::transforms::Identifier;
 use vegafusion_core::task_graph::task_value::TaskValue;
+use vegafusion_dataframe::dataframe::DataFrame;
 
 #[async_trait]
 impl TransformTrait for Identifier {
