+++ vegafusion-runtime/src/transform/impute.rs
@@ -1,5 +1,5 @@
 use crate::expression::compiler::config::CompilationConfig;
-use crate::sql::dataframe::DataFrame;
+
 use crate::transform::TransformTrait;
 use async_trait::async_trait;
 use datafusion::common::ScalarValue;
@@ -11,6 +11,7 @@ use vegafusion_common::error::{Result, ResultWithContext};
 use vegafusion_common::escape::unescape_field;
 use vegafusion_core::proto::gen::transforms::Impute;
 use vegafusion_core::task_graph::task_value::TaskValue;
+use vegafusion_dataframe::dataframe::DataFrame;
 
 #[async_trait]
 impl TransformTrait for Impute {
