+++ vegafusion-runtime/src/transform/joinaggregate.rs
@@ -1,5 +1,4 @@
 use crate::expression::compiler::config::CompilationConfig;
-use crate::sql::dataframe::DataFrame;
 use crate::transform::aggregate::make_aggr_expr_for_named_col;
 use crate::transform::TransformTrait;
 use async_trait::async_trait;
@@ -10,6 +9,7 @@ use vegafusion_core::error::Result;
 use vegafusion_core::proto::gen::transforms::{AggregateOp, JoinAggregate};
 use vegafusion_core::task_graph::task_value::TaskValue;
 use vegafusion_core::transform::aggregate::op_name;
+use vegafusion_dataframe::dataframe::DataFrame;
 
 #[async_trait]
 impl TransformTrait for JoinAggregate {
