+++ vegafusion-runtime/tests/util/check.rs
@@ -14,15 +14,14 @@ use vegafusion_common::error::Result;
 use vegafusion_core::expression::parser::parse;
 use vegafusion_core::proto::gen::transforms::TransformPipeline;
 use vegafusion_core::spec::transform::TransformSpec;
-use vegafusion_runtime::expression::compiler::call::make_session_context;
+use vegafusion_dataframe::connection::Connection;
 use vegafusion_runtime::expression::compiler::compile;
 use vegafusion_runtime::expression::compiler::config::CompilationConfig;
 use vegafusion_runtime::expression::compiler::utils::ExprHelpers;
-use vegafusion_runtime::sql::connection::datafusion_conn::DataFusionConnection;
-use vegafusion_runtime::sql::connection::Connection;
 use vegafusion_runtime::task_graph::timezone::RuntimeTzConfig;
 use vegafusion_runtime::tokio_runtime::TOKIO_RUNTIME;
 use vegafusion_runtime::transform::pipeline::TransformPipelineUtils;
+use vegafusion_sql::connection::datafusion_conn::{DataFusionConnection, make_datafusion_context};
 
 pub fn check_expr_supported(expr_str: &str) {
     let expr = parse(expr_str).unwrap();
@@ -117,7 +116,7 @@ pub fn eval_vegafusion_transforms(
     transform_specs: &[TransformSpec],
     compilation_config: &CompilationConfig,
 ) -> (VegaFusionTable, Vec<ScalarValue>) {
-    let ctx = make_session_context();
+    let ctx = make_datafusion_context();
     let conn = Arc::new(DataFusionConnection::new(Arc::new(ctx))) as Arc<dyn Connection>;
 
     // add ordering column
