+++ vegafusion-server/Cargo.toml
@@ -26,6 +26,10 @@ predicates = "2.1"
 tonic-build = "0.8.4"
 prost-build = "0.11.4"
 
+[dependencies.vegafusion-common]
+path = "../vegafusion-common"
+version = "1.0.3"
+
 [dependencies.vegafusion-core]
 path = "../vegafusion-core"
 features = [ "tonic_support",]
